gave up on fermiumbooter dependency
reasons:
- some of those modded ASM target code loaded too early for mixins to be usable
- im kinda overwhelmed with the indepth ASM so im having a hard time translating it.

now mixins is just shaded in, as its supposed to be for vanilla targeting mixins.

i see that LLibraryHooks ASM doesnt work, maybe due to new mappings? otherwise it seems fine

fermiumbooter is still required in the build.gradle dependencies in order for runClient to work. this would also work with MixinBootstrap or other mixin loading dependencies and doesnt change the output jar

note for the deletion of the coremodcontainer class: this is not needed for what youre doing :)

another note: for some reason this sometimes doesn't generate refmaps. if your jar doesnt contain a refmap,advcreation.json file in main folder, you have to delete your build folder and recompile